+++ src/main/java/de/superrental/carrentalinterface/repository/RentalRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface RentalRepository extends CrudRepository<Rental, Long> {
-    List<Rental> findByCustomer_customerId(Long customerId);
+public interface RentalRepository extends CrudRepository<Rental, Integer> {
+    List<Rental> findByCustomer_customerId(Integer customerId);
 }
